requirements.txt
redis
aiofiles
aiosqlite
discord.py
psutil
python-dotenv
requests
rgbprint
sentry-sdk